add new rules file for basic bowtie2 mapping
This is Jesse's Snakefile for bowtie2 alignments and automated IGV
alignment images, edited to use Sunbeam's configuration variables.  As
written it assumes the input samples are a single pair of fasta files
and only does one alignment per genome segment, so it needs to be
updated to handle multiple inputs and show those in each IGV image.