Fix #58: Add multithreading support for mapping rules (#60)
* add new rules file for basic bowtie2 mapping

This is Jesse's Snakefile for bowtie2 alignments and automated IGV
alignment images, edited to use Sunbeam's configuration variables.  As
written it assumes the input samples are a single pair of fasta files
and only does one alignment per genome segment, so it needs to be
updated to handle multiple inputs and show those in each IGV image.

* generalize mapping for arbitrary number of samples

This updates the bowtie2/samtools/bcftools rules to handle more than one
input sample.  Now each input sample will be mapped to each input genome
segment.  This breaks the IGV step as that still needs to be updated to
handle multiple samples per genome segment, though.

* generalize IGV function for multiple input samples

This updates the igv_snapshot rule to use all sorted bam files as input,
and makes igv_render_script() accept a list of bam files and render all
of them together in one IGV image per genome segment.

* generalize the socket-based IGV functions too

* add non temp test dir option

* assume one genome per index fasta file for mapping

* fix typo in bowtie.rules

* use MAPPING_FP for bowtie2_build

* use MAPPING_FP for all mapping rules

* fix alignment image target strings for mapping

* remove extra print statement

* move TARGET_MAPPING setup from mapping to targets

This dumps an ugly block of code straight into targets.rules (so it can
be used in both TARGET_ALL and inside bowtie.rules) but it should
probably go elsewhere.

* add mapping targets and configuration for tests

* bwa -> bowtie2 in requirements.txt (for mapping)

* adding bwa back to requirements.txt (in use by qc)

Whoops, even though bwa isn't called directly in the sunbeam rules, it's
our default method in the config file for the qc step (to be used by
decontaminate.py installed alongside sunbeam).

* add simple IGV install and use in tests

* install IGV into $HOME instead of current dir

* downgrade installed IGV from 2.3.92 to 2.3.68

The latest IGV version generates blank image panes when run with a batch
script, so this rolls back to a known working version.

* integrate mapping rules with sunbeamlib

This adds a utility function to sunbeamlib to help identify sequences in
genome files, and uses an existing utility function for the lists of
bowtie2 index files.  This cleans up targets.rules and better follows
existing conventions.

* install IGV in local/IGV and define in config file

Instead of putting a local copy of IGV in $HOME, it will go into the
local subdirectory of the sunbeam working directory.  A new config
option Cfg['mapping']['igv_fp'] defines the executable to run.

* update and reorganize both IGV rendering functions

Now there's just one igv_render() that calls either the script- or
socket-based IGV-control function as needed, removing some duplicate
code and making both use the same IGV commands.

* log conda environment setup to file for Travis

Travis jobs fail if more than 4 MB of text appears on standard output.
This hides most of the text from install.sh to keep below that limit.

* install xdotool for Travis

* Use BioPython's SeqIO.parse() to read record IDs

This replaces a custom fasta-parsing function with one already available
from BioPython.

* Move IGV python code from rules to sunbeamlib

This puts the IGV-control functions into a separate IGV module in
sunbeamlib, and just imports the one public function into the mapping
rules file.

* Fix #58: add threads option for mapping rules

This adds a threads configuration option for the mapping section, and
uses that value for the Snakemake "threads" variable for bowtie2 and
samtools rules.