Fix #59: Support custom IGV preferences
This adds an igv_prefs option to the mapping configuration, which is
passed to the IGV rendering code and used to override specific IGV
preferences.  The default NAME_PANEL_WIDTH setting should now work
better for long input filenames.  The filenames are also now sorted when
given to IGV so the display order is consistent.