Fix use of mutable value for function default
This uses None as the default for igv_prefs in render() instead of {}.
Landscape pointed out that dictionary will only be instantiated when the
function definition is processed, so all calls to render() will use the
same shared dictionary object.  Using None and then handling the
assignment inside the function is safer.