Add IGV preferences support for mapping rules (#61)
* Fix #59: Support custom IGV preferences

This adds an igv_prefs option to the mapping configuration, which is
passed to the IGV rendering code and used to override specific IGV
preferences.  The default NAME_PANEL_WIDTH setting should now work
better for long input filenames.  The filenames are also now sorted when
given to IGV so the display order is consistent.

* merge mapping-multithreaded

* Fix use of mutable value for function default

This uses None as the default for igv_prefs in render() instead of {}.
Landscape pointed out that dictionary will only be instantiated when the
function definition is processed, so all calls to render() will use the
same shared dictionary object.  Using None and then handling the
assignment inside the function is safer.