fix #39: bypass cap3 for empty contigs
This adds a check for an empty input file for the "cap3" rule.  In that
case the file is just passed along to the output, bypassing cap3.  The
subsequent rules "save_intermediates" and "final_filter" are able to
handle empty files already, so the assembly step should now run to
completion even for samples where no contigs can be built.