Fix #39: Bypass cap3 for empty contigs in assembly rules (#76)
* add failing test to address #39

This adds a third dummy fastq input pair, using arbitrary reads so that
no contigs are build by IDBA-UD.  cap3 then fails with a segmentation
fault as described in the issue.

* add more test targets for new test sample

* fix #39: bypass cap3 for empty contigs

This adds a check for an empty input file for the "cap3" rule.  In that
case the file is just passed along to the output, bypassing cap3.  The
subsequent rules "save_intermediates" and "final_filter" are able to
handle empty files already, so the assembly step should now run to
completion even for samples where no contigs can be built.

* fix igv input handling for snakemake 3.11.1

The latest snakemake provides objects of type snakemake.io.Namedlist for
some that previously were strings, and igv.render() can't cast those to
Path objects as-is.  This converts them to str explicitly.

* add apt-get update for travis