add failing test for #153
Add a new function to the test suite, test_blank_fp_behavior, that
checks to make sure output files are *not* present when genomes_fp is a
blank string in the config (even if fasta files are at the top level).
This should fail until some extra logic is added to the Snakefile and
config file loader.